fix entitiy

+++ api/src/main/java/ntk/android/base/entitymodel/blog/BlogContentModel.java
@@ -37,7 +37,7 @@ public class BlogContentModel extends BaseModuleEntity<Long> {
     @SerializedName("scoreSumPercent")
     public long scoreSumPercent;
     @SerializedName("viewCount")
-    public long viewCount;
+    public int viewCount;
     @SerializedName("favorited")
     public boolean favorited;
     @SerializedName("mainImageSrc")
